requirements.txt
# requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
gunicorn==21.2.0
pydantic==2.10.3
pinecone-client==5.0.1
sentence-transformers==3.2.1
requests==2.31.0
python-dotenv==1.0.0